microservices/auth-service/requirements.txt
uvicorn
fastapi
pydantic
mysql-connector-python